# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.editUserRequest:
    properties:
      name:
        example: John Doe
        type: string
      position:
        example: CTO
        type: string
    required:
    - name
    - position
    type: object
  handlers.failedResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.userMeResponse:
    properties:
      email:
        example: alice@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Alice
        type: string
      position:
        example: CEO
        type: string
    type: object
...
  handlers.vcModel:
    properties:
      approved:
        example: false
        type: boolean
      email:
        example: example@vnest.org
        type: string
      id:
        example: 1
        type: integer
      name:
        example: someone
        type: string
    type: object
info:
  contact: {}
  description: This endpoint is for dev purposes
  title: VNEST Dashboard Swagger docs
  version: "1.0"
paths:
...
  /manage/vc/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the VC from the database
      parameters:
      - description: VC ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      security:
      - BearerAuth: []
      summary: Delete a VC
      tags:
      - admin
  /manage/vc/{id}/approve:
    put:
      consumes:
      - application/json
      description: Sets the VC's approved field to true
      parameters:
      - description: VC ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      security:
      - BearerAuth: []
      summary: Approve a VC
      tags:
      - admin
  /manage/vc/{id}/remove:
    put:
      consumes:
      - application/json
      description: Sets the VC's approved field to false
      parameters:
      - description: VC ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.failedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      security:
      - BearerAuth: []
      summary: Unapprove a VC
      tags:
      - admin
  /manage/vc/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of users with the VC role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.vcModel'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.failedResponse'
      security:
      - BearerAuth: []
      summary: Get list of VCs
      tags:
      - admin
...
  /user/delete:
    delete:
      description: Deletes the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - user
  /user/edit:
    put:
      consumes:
      - application/json
      description: Allows a user to edit only their name and position
      parameters:
      - description: Edit user name and position
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.editUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit user profile
      tags:
      - user
  /user/me:
    get:
      description: Returns details of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userMeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" and your JWT token for authentication and authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"